Fix undefined behavior in shader fog handling
`smoothstep` is only defined when `edge0 < x < edge1`, but we were
reversing the two edges incorrectly, which results in undefined behavior.

Additionally, this fixes a problem where the alpha component of the fog
color was ignored. This is used in vanilla to apply a "weight" to the
final blend factor prior to the fog color being mixed in with the
fragment color.