Don't show donation popup with Controllable installed
The mechanism used to display the prompt appears to conflict with
Controllable's cursor, causing it to vanish and also not interact
with any GUI elements.

There isn't an obvious fix after a few minutes of debugging, and
since we plan to rewrite the whole GUI in the near future, it's
not worth debugging further.

It's better to miss out a few potential donations than actively jeopardize
gameplay.

Related: #248