Do not expand client chunk cache array beyond vanilla size
The vanilla implementation relies on collisions with old data at
the same index in order for chunks to correctly unload in all cases
(see ClientChunkCache.Storage.replace when it calls getAndSet).
Because the array is larger due to the power-of-two trick,
the collision never happens, which keeps some data about the
chunk present in the level (e.g. block entity tickers).

An upstream Sodium bug report highlighted that in particular
this causes campfires to emit particles even when far away.

This might also fix an issue with Create contraptions rendering
off in the distance.