Apply smooth lighting to any fluid that doesn't emit light, not just water
This makes the logic more consistent with that of block rendering,
and allows modded fluids to benefit from this visual improvement.
Hopefully, none are depending on the visuals of flat lighting.

Lava continues to render as before (no smooth lighting) as it has a light
level of 15.

Thanks to MehVahdJukaar for suggesting the change.